Moved localstorage functions to a single file
