--- utils/localStorage/getLocalStorage.ts
@@ -1,9 +1,0 @@
-/**
- *
- * @param key keyID in localStorage
- *
- * @returns a string that contains the data of specific keyID in localStorage
- */
-export function getLocalStorage(key: string) {
-  return window.localStorage.getItem(key);
-}
